test(desktop): isolate entry routing helper
